MAINT Merge math_util and math_util_test
This is to reflect the merging of these modules in the source.
